+++ main.go
@@ -6,7 +6,7 @@ import (
 )
 
 var (
-	version = "v0.5.1"
+	version = "v0.5.2"
 	commit  = "unknown"
 )